+++ src/components/ids-data-grid/demos/virtual-scroll-infinite-scroll.ts
@@ -92,11 +92,11 @@ const setData = async () => {
 
 setData();
 
-dataGrid.addEventListener('virtualscroll-top', async (e: Event) => {
+dataGrid.addEventListener('virtualscrolltop', async (e: Event) => {
   console.info(`Virtual Scroll reached top`, (<CustomEvent>e).detail);
 });
 
-dataGrid.addEventListener('virtualscroll-bottom', async (e: Event) => {
+dataGrid.addEventListener('virtualscrollbottom', async (e: Event) => {
   console.info(`Virtual Scroll reached bottom`, (<CustomEvent>e).detail);
   const MAX_ROW_INDEX = 2999;