+++ doc/CHANGELOG.md
@@ -7,7 +7,7 @@
 - `[Alert]` Added the ability to use any icon and set any alert color on the alerts. ([#1138](https://github.com/infor-design/enterprise-wc/issues/1138))
 - `[Badge]` Fixed uneven shape on badge icons. ([#1014](https://github.com/infor-design/enterprise-wc/issues/1014))
 - `[Build]` Updated `esbuild` to fix issues with import order. ([#1140](https://github.com/infor-design/enterprise-wc/issues/1140))
-- `[DataGrid]` Add events for `virtualscroll-top` and `virtualscroll-bottom`. ([#1102](https://github.com/infor-design/enterprise-wc/issues/1102))
+- `[DataGrid]` Add events for `virtualscrolltop` and `virtualscrollbottom`. ([#1102](https://github.com/infor-design/enterprise-wc/issues/1102))
 - `[Icons]` Changed the way custom icons work so they can be used only at one time and from a file. ([#1122](https://github.com/infor-design/enterprise-wc/issues/1122))
 - `[Icons]` Clean up examples for icons. ([#509](https://github.com/infor-design/enterprise-wc/issues/509))
 - `[Locale]` Locale information files and messages are now separate from the build. They must be served as assets from the `node_modules/ids-enterprise-wc/locale-data` folder. ([#1107](https://github.com/infor-design/enterprise-wc/issues/1107))